Bug fix when removing missing data when using axis.All()
When extracting multiple fields from data, missing values must must be
inserted in all places where 1 or more fields have a missing value. The
code that ensures this wrote missing values into the wrong in the 3D
data arrays.

Missing values were inserted using the date-index only (not leadtime and
location), which was treated as a linear index into the flattened
array. This means that extra missing values were inserted likely on the
first few days of the dataset.

This affects only -m timeseries and newer diagrams such as -m autocorr
and -m fss, which extract data using axis.All(). The effect is that in
some cases, some valid forecast/observation data is assigned missing
when they shouldn't be.