Fix bug in -m within for axis values that are empty
When an axis value (f.ex a date) has no values, then within always
showed a 0. This is because there was a bug in verif.interval, where an
input value of nan always resulted in a 0. Luckily, in most cases
missing values are filtered out before being sent to verif.interval.
However, in -m within this does not happen when an axis value has no
elements.