Fixed bug when reading text files with more than 256 columns
An error occurred when there are more than 256 columns in a text file,
since the comparison of the number of columns on each line fails:

len(row) is not len(header)

This does not work for larger numbers, and "is not" is now replaced with
"==".

Thanks to @Ciccuz for reporting.